+++ CubeEngine/EngineSrc/Event/EventMgr.h
@@ -5,7 +5,9 @@
 #include <string>
 #include "../EngineSrc/Math/vec2.h"
 #include "../Engine/EngineDef.h"
+#include <unordered_map>
 namespace tzw {
+class Node;
 struct EventInfo
 {
     int type;
@@ -13,24 +15,38 @@ struct EventInfo
     int arg;
     vec2 pos;
 };
-class Event;
+class EventListener;
 class EventMgr
 {
 public:
-    TZW_SINGLETON_DECL(EventMgr)
-    void addEventListener(Event * event);
+    void addListener(EventListener * listener);
+    void addListener(EventListener * listener, Node * node);
+    void addFixedPiorityListener(EventListener * event);
+    void addNodePiorityListener(Node * node,EventListener * event);
+
     void handleKeyPress(std::string keyCode);
     void handleKeyRelease(std::string keyCode);
     void handleMousePress(int button,vec2 pos);
     void handleMouseRelease(int button,vec2 pos);
     void handleMouseMove(vec2 pos);
     void apply(float delta);
-    void removeEventListener(Event * event);
-    void sortEvents();
+    void removeEventListener(EventListener * event);
+    void sortFixedListener();
 private:
-    EventMgr();
-    std::vector<Event *>m_list;
+    void sortNodePiorityListener();
+    void visitNode(Node * node);
+    void applyKeyPress(EventInfo & info);
+    void applyKeyRelease(EventInfo & info);
+    void applyMousePress(EventInfo & info);
+    void applyMouseRelease(EventInfo & info);
+    void applyMouseMove(EventInfo & info);
+private:
+    std::vector<EventListener *>m_list;
+    std::deque<EventListener *>m_NodePioritylist;
     std::deque<EventInfo>m_eventDeque;
+    std::unordered_map<Node *, EventListener *> m_nodeListenerMap;
+
+    TZW_SINGLETON_DECL(EventMgr)
 };
 
 } // namespace tzw